Add files via upload
Personal Loan Campaign - Problem Statement
Description
Background and Context
AllLife Bank is a US bank that has a growing customer base. The majority of these customers are liability customers (depositors) with varying sizes of deposits. The number of customers who are also borrowers (asset customers) is quite small, and the bank is interested in expanding this base rapidly to bring in more loan business and in the process, earn more through the interest on loans. In particular, the management wants to explore ways of converting its liability customers to personal loan customers (while retaining them as depositors).
A campaign that the bank ran last year for liability customers showed a healthy conversion rate of over 9% success. This has encouraged the retail marketing department to devise campaigns with better target marketing to increase the success ratio.
You as a Data scientist at AllLife bank have to build a model that will help the marketing department to identify the potential customers who have a higher probability of purchasing the loan.
Objective
To predict whether a liability customer will buy a personal loan or not.
Which variables are most significant.
Which segment of customers should be targeted more.
Conclusion:
1. Comparing the models gotten through Logistic Regression with the ones of Decision tree, it still appears Decision Tree with post-pruning model will work best for the project objective as the wrong classification of potential personal loan customer is brought to the minimum level with this model.
2. Education, Income and Family Size are the most significant variables for identifying a personal loan customer
Summary of Insights and Recommendations
1. Customers with family size 3 and 4 have higher conversion rate to asset customers than others. Marketing campaign on loan should therefore be focused on this class of customers
2. Customers with at least Graduate Education have higher conversion rate to asset customers. This must be considered in the next loan campaign of the bank to have higher success rate
3. Customers without securities account are 7 times more likely become asset customer than those without securities account. All Life Bank marketing team needs to focus on these customers
4. Customers with internet banking facilities have slightly higher conversion rate to asset customers than others. However, I will advise the marketing team ensures their next loan campaign reaches all customers whether they have internet banking facilities or not. They can create a personal loan product accessible on the bank's internet banking platform(this will attract internet banking customers) and other platforms accessbile to non-internet banking customers
5. Customers without other bank's credit card take personal loan more in All Life Bank than others with other bank's credit card. However, conversion to asset customer rates for customers with and without other bank's credit card are not significantly different. Marketing team should therefore take their campaigns to all customers
6. Customers from Los Angeles, Santa Clara, San Diego and Alameda constitute over 59% of all the bank's personal loan customers. The marketing team should therefore focus their marketing campaigns to these counties as they also constitute 55% of their total customer base. However, worthy of mention is that customers from Sonoma county has the highest conversion to loan customer rate as approximately 3 in every 10 customers from that county are likely to take loan. Marketing effort to drive onboarding more customers from Sonoma County is hereby recommended.