Add files via upload
Travel Package Purchase Prediction - Problem Statement
Description
Background and Context
You are a Data Scientist for a tourism company named "Visit with us". The Policy Maker of the company wants to enable and establish a viable business model to expand the customer base.

A viable business model is a central concept that helps you to understand the existing ways of doing the business and how to change the ways for the benefit of the tourism sector.

One of the ways to expand the customer base is to introduce a new offering of packages.

Currently, there are 5 types of packages the company is offering - Basic, Standard, Deluxe, Super Deluxe, King. Looking at the data of the last year, we observed that 18% of the customers purchased the packages.

However, the marketing cost was quite high because customers were contacted at random without looking at the available information.

The company is now planning to launch a new product i.e. Wellness Tourism Package. Wellness Tourism is defined as Travel that allows the traveler to maintain, enhance or kick-start a healthy lifestyle, and support or increase one's sense of well-being.

However, this time company wants to harness the available data of existing and potential customers to make the marketing expenditure more efficient.

You as a Data Scientist at "Visit with us" travel company have to analyze the customers' data and information to provide recommendations to the Policy Maker and Marketing Team and also build a model to predict the potential customer who is going to purchase the newly introduced travel package.

Objective
To predict which customer is more likely to purchase the newly introduced travel package.

Summary of Insights and Recommendations¶
Most of the company customers that purchased travel package last year are less than 38 years. It is therefore recommended that the marketing campaign of the new product should be focused on customers within that age limit.

Over 66% of the travel packages sold last time were purchased by self enquiry customers. However, the company invited customers have higher likelihood of purchasing travel package than self enquiry customers. The marketing team is hereby advised that the two channels should be explored to the fullest in pitching the new products so as to make better sales.

Customers from Tiers 1 and 3 cities constitute 95% of the travel package purchasers at Visit-with-us. Most of the marketing efforts on the new product should therefore be concentrated in those two cities tiers.

Over 86% of those customers that purchase travel packages are salary earners and small business owners. It is therefore recommended that the new product should be marketed more to people with such occupations.

Most of the customers are males. The marketing team of the company should therefore take sales pitches with males more seriously.

Over 68% are either Single or Married. Marketing team should focus more of their sales pitches on customers with these marital statuses.

The customers with passport are far more likely to purchase travel packages than those without passport. The company should therefore focus thier sales pitches more on passport holding customers. The company may also begin to render passport processing facilitation services for their customers at a fee. This will serve as a complimentary service that will assist to drive sales of travel packages.

Customers who own cars are far more likely to purchase travel packages than those non-car owners. The marketing team should consider this in pitching their new product.

The bulk of the customers purchasing travel packages are either of Executive Status or Manager Status. These sets of customers are to be focused when pitching the new product.

Basic and Deluxe are the two products accounting for over 82% of the company sales last year. The new product to be introduced should be carefully designed to have both Basic and Deluxe variants so as to make good sales.

Customers who had at least 5 follow-ups from the company salespersons are more likely to purchase travel packages than other customers. This sales system should be adopted in the process of pitching the new product.

Most of those that purchase travel packages earn below 24,000 monthly. The company should therefore consider this when pitching the new product.